EMI: Rename EMI's sound-system to emisound
And move it into a folder of it's own.
